+++ Solution/RiskOfChaos/Content/RoCContent.cs
@@ -84,6 +84,10 @@ namespace RiskOfChaos.Content
                     }
                 }
 
+#if DEBUG
+                Log.Debug($"Collecting content from initializer {i + 1}/{contentInitializerMethods.Count} ({contentInitializerMethod.FullDescription()})");
+#endif
+
                 object returnValue = contentInitializerMethod.Invoke(null, parameters);
 
                 IEnumerator enumerator = null;