+++ Solution/RiskOfChaos/EffectDefinitions/World/EnableRandomArtifact.cs
@@ -6,6 +6,7 @@ using RiskOfChaos.EffectHandling.EffectClassAttributes;
 using RiskOfChaos.EffectHandling.EffectClassAttributes.Data;
 using RiskOfChaos.EffectHandling.EffectClassAttributes.Methods;
 using RiskOfChaos.EffectHandling.EffectComponents;
+using RiskOfChaos.EffectHandling.Formatting;
 using RiskOfChaos.SaveHandling;
 using RiskOfChaos.Utilities.Extensions;
 using RiskOfOptions.OptionConfigs;
@@ -13,11 +14,12 @@ using RoR2;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using System.Text;
 using UnityEngine.Networking;
 
 namespace RiskOfChaos.EffectDefinitions.World
 {
-    [ChaosTimedEffect("enable_random_artifact", TimedEffectType.UntilStageEnd)]
+    [ChaosTimedEffect("enable_random_artifact", TimedEffectType.UntilStageEnd, HideFromEffectsListWhenPermanent = true)]
     [EffectConfigBackwardsCompatibility("Effect: Enable Random Artifact (Lasts 1 Stage)")]
     public sealed class EnableRandomArtifact : NetworkBehaviour
     {
@@ -107,6 +109,7 @@ namespace RiskOfChaos.EffectDefinitions.World
         }
 
         ChaosEffectComponent _effectComponent;
+        ChaosEffectNameComponent _effectNameComponent;
 
         [SerializedMember("a")]
         ArtifactIndex _enabledArtifact = ArtifactIndex.None;
@@ -114,6 +117,7 @@ namespace RiskOfChaos.EffectDefinitions.World
         void Awake()
         {
             _effectComponent = GetComponent<ChaosEffectComponent>();
+            _effectNameComponent = GetComponent<ChaosEffectNameComponent>();
         }
 
         public override void OnStartServer()
@@ -141,6 +145,11 @@ namespace RiskOfChaos.EffectDefinitions.World
             if (NetworkServer.active && _enabledArtifact != ArtifactIndex.None)
             {
                 RunArtifactManager.instance.SetArtifactEnabledServer(ArtifactCatalog.GetArtifactDef(_enabledArtifact), true);
+
+                if (_effectNameComponent)
+                {
+                    _effectNameComponent.SetCustomNameFormatter(new NameFormatter(_enabledArtifact));
+                }
             }
         }
 
@@ -149,7 +158,75 @@ namespace RiskOfChaos.EffectDefinitions.World
             if (!NetworkServer.active || !RunArtifactManager.instance)
                 return;
 
-            RunArtifactManager.instance.SetArtifactEnabledServer(ArtifactCatalog.GetArtifactDef(_enabledArtifact), false);
+            if (_enabledArtifact != ArtifactIndex.None)
+            {
+                RunArtifactManager.instance.SetArtifactEnabledServer(ArtifactCatalog.GetArtifactDef(_enabledArtifact), false);
+            }
+        }
+
+        class NameFormatter : EffectNameFormatter
+        {
+            ArtifactIndex _enabledArtifactIndex;
+
+            public NameFormatter(ArtifactIndex enabledArtifactIndex)
+            {
+                _enabledArtifactIndex = enabledArtifactIndex;
+            }
+
+            public NameFormatter()
+            {
+            }
+
+            public override string GetEffectNameSubtitle(ChaosEffectInfo effectInfo)
+            {
+                string subtitle = base.GetEffectNameSubtitle(effectInfo);
+
+                ArtifactDef artifactDef = ArtifactCatalog.GetArtifactDef(_enabledArtifactIndex);
+                if (artifactDef)
+                {
+                    StringBuilder stringBuilder = HG.StringBuilderPool.RentStringBuilder();
+                    if (!string.IsNullOrWhiteSpace(subtitle))
+                    {
+                        stringBuilder.AppendLine(subtitle);
+                    }
+
+                    stringBuilder.Append("\n(");
+
+                    stringBuilder.Append("<style=cArtifact>");
+                    stringBuilder.Append(Language.GetString(artifactDef.nameToken));
+                    stringBuilder.Append("</style>");
+
+                    stringBuilder.Append(")");
+
+                    subtitle = stringBuilder.ToString();
+
+                    stringBuilder = HG.StringBuilderPool.ReturnStringBuilder(stringBuilder);
+                }
+
+                return subtitle;
+            }
+
+            public override object[] GetFormatArgs()
+            {
+                return [];
+            }
+
+            public override void Serialize(NetworkWriter writer)
+            {
+                writer.WritePackedIndex32((int)_enabledArtifactIndex);
+            }
+
+            public override void Deserialize(NetworkReader reader)
+            {
+                _enabledArtifactIndex = (ArtifactIndex)reader.ReadPackedIndex32();
+                invokeFormatterDirty();
+            }
+
+            public override bool Equals(EffectNameFormatter other)
+            {
+                return other is NameFormatter otherFormatter &&
+                       _enabledArtifactIndex == otherFormatter._enabledArtifactIndex;
+            }
         }
     }
 }